+++ lib/std/debug.lua
@@ -185,26 +185,6 @@ if _DEBUG.argcheck then
   end
 
 
-  --- Ordered iterator for integer keyed values.
-  -- Like ipairs, but does not stop at the first nil value.
-  -- @tparam table t a table
-  -- @treturn function iterator function
-  -- @treturn table t
-  -- @usage
-  -- for i,v in argpairs {"one", nil, "three"} do print (i, v) end
-  local function argpairs (t)
-    local i, max = 0, 0
-    for k in pairs (t) do
-      if type (k) == "number" and k > max then max = k end
-    end
-    return function (t)
-	    i = i + 1
-	    if i <= max then return i, t[i] end
-	   end,
-    t, true
-  end
-
-
   --- Strip trailing ellipsis from final argument if any, storing maximum
   -- number of values that can be matched directly in `t.maxvalues`.
   -- @tparam table t table to act on